feat: doing Order(commande gestion) entity

+++ src/entity/Dish.java
@@ -58,7 +58,7 @@ public class Dish {
 
         for (IngredientWithQuantity ingredient : getIngredients()) {
             resultList.add(
-                    ingredient.getIngredient().getAvalaibleQuantity().getQuantity() / ingredient.getRequiredQuantity()
+                    Double.valueOf(ingredient.getIngredient().getAvalaibleQuantity().getQuantity() / ingredient.getRequiredQuantity())
             );
         }
         Collections.sort(resultList);
@@ -71,15 +71,13 @@ public class Dish {
 
         for (IngredientWithQuantity ingredient : getIngredients()) {
             resultList.add(
-                    ingredient.getIngredient().getAvalaibleQuantity(date).getQuantity() / ingredient.getRequiredQuantity()
+                    Double.valueOf(ingredient.getIngredient().getAvalaibleQuantity(date).getQuantity() / ingredient.getRequiredQuantity())
             );
         }
         Collections.sort(resultList);
         if(resultList.isEmpty()) return 0.0;
         return resultList.get(0);
     }
-
-
 
     @Override
     public boolean equals(Object o) {
